more splitting and cleanup of rag
